Introduce basic tagging infrastructure. r=jan
From Jan's review:

> this 'all-in-one-go' change of tagsets in the database looks nice as code,
> but from a UI perspective you'd want each change (addition, deletion, size
> change) to be saved right away. Think about crappy internet connections -
> you want each change to be stored in order not to loose work.

My thinking was that we'd just send the whole new tagset on any change (or
incrementally every 5 seconds, or something). The underlying HTTP request
has a fair amount of overhead itself, and I'm imagining a typical tagset
size being something like 6. So I'm not sure that we'd be gaining all that
much by sending incremental updates, and it adds code complexity.

We should definitely keep that in mind as a future optimization, though.