Switch to 4-space indentation, using python's reindent, to comply with PEP 8. v1 r=me
This is automated and verifiable enough that it doesn't require manual review.
In particular, 'git diff -w', which gives the diff disregarding whitespace changes,
returns empty for this patch. This doesn't mean it's entirely safe, since it can
mess up wrapping (and since whitespace isn't neutral in python), but I trust the
tool not to do anything too sinister.